feat(stepfunctions): use readonly mode of workflowStudio to render StateMachine in CDK Applications (#6874)
## Problem
As we added support for WorkflowStudio in Toolkit for editing ASL
language files, we want to upgrade the render StateMachine Graph in CDK
applications experience with the a readonly mode of WorkflowStudio which
has better visualization, a code pane for view ASL definition, an export
mechanism for exporting ASL definition in JSON/YAML or PNG/SVG.

## Solution
Removed the old visualization classes that links to the legacy renderer
CDN. From the `renderStateMachineGraphCDK` wizard, upon finding the
`template.json` file for a specific stateMachine, we send the file uri
to WorkflowStudioEditorProvider to open a readonly WFS to render the
StateMachine Graph.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.