fix(amazonq): inline chat now activates independently of LSP chat #6973
## Problem
When toggling the experiment via, the `aws.experiments.amazonqChatLSP`
flag in settings, inline chat fails to activate. Specifically, it
appears the command fails to register.
```
command 'aws.amazonq.inline.invokeChat' not found
```

This happened because inline chat was activated as part of regular chat.
When we use LSP chat, this then never gets activated.

## Solution
- Activate inline chat independently of regular chat. 

## Testing 
- Verified the fix by running extension with and without
`aws.experiments.amazonqChatLSP` flag and using inlineChat in both
cases. One of these cases is shown in the demo below.

## Notes
This will have to change once we consume LSP inline chat, as we don't
want to activate both our inline chat and Flare's.