fix(amazonq): validate customization on profile change #7049
## Problem
- Before region expansion:
  customization is bound to a specific idc instance
- After region expansion:
  customization is bound to a specific Q profile and an idc instance can
  have multi profiles
  - therefore each Q profile will have access to different sets of
    customization

## Solution
We need to validate if the selected customization is accessible from the
user's selected profile, otherwise will get
```
*An error occurred while processing your request.*
This error is reported to the team automatically. We will attempt to fix it as soon as possible.
Details: The provided profile ARN and customization ARN is mismatched. (Service: CodeWhispererRuntime, Status Code: 403, 
```