feat(core): Store codewhisperer customization in global state as single value and consume new authUtil (#7067)
## Problem
With the old VSCode auth, customization was stored in global state
(indexed by connection ID). With the new auth from the Flare identity
server, this no longer applies since there is only a single connection

## Solution
* Consume the new AuthUtil in the customizationUtil
* Change `CODEWHISPERER_SELECTED_CUSTOMIZATION` in the global state to
be a single value instead of a map

## Follow-up
Investigate if it's possible to hook up the customization to the Q
profile from the new Region Manager, so customizations are persisted
between different logins

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.