fix customizations change not being synced to lsp
Problem:

A message was already set up in a previous commit to send the lsp a
message when the Q customization model was changed, but it didn't work
as expected.

Solution:

There are different lsp messages sent to indicate a change in user settings.
For auth Q profile it is one type of Request, and for the customization selection,
it uses a different lsp notification.

Each type of message has its own handler in the language server, so if we send the
message using the wrong message type, it will not be handled correctly since the correct
handler is somewhere else.

The notable handling functions were:
- getAmazonQRelatedWorkspaceConfigs(), where it is part of the handler for `didChangeConfiguration()`
- setupConfigurationListeners(), where it is part of the handler for `onUpdateConfiguration()`

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>