fix(amazonq): Sync client settings to language server (#7114)
When certain settings/configs are changed in the IDE, they need to be
propagated to the language server.
The settings were the customizationArn, the users opt in preference for
telemetry, and some other things.

This PR:
- Pushes the new customization arn when it is changed. This arn only
exists when the user is using the non-default customization, otherwise
it is undefined.
- A small hack needed to be made to explicitly push this change, since
by default we utilized the automatic push to the LS when the users
settings file changed (eg opt in telemetry). But because customization
is not stored in the users settings, we didn't get that for free. See
the command `aws.amazonq.updateCustomizations` which was made to handle
this edge case. Which eventually **explicitly** uses the language server
message `DidChangeConfigurationNotification`
- Pushes the opt out telemetry setting to the language server
- Since this setting is stored in the vscode settings json, the push to
language server mechanism comes for free under the
`DidChangeConfigurationNotification` message.
- The Q Auth Profile ARN is a separate config that needed to be handled
by a different handler (different from customization), so the function
`pushConfigUpdate()` was created to route the correct config to the
correct handler in the LS
- This uses the message `updateConfigurationRequestType` which is
different from `DidChangeConfigurationNotification` that the other
configs use. This results in a different code path being hit, which is
expected.

Some helpful notes:
- `getAmazonQRelatedWorkspaceConfigs()` (triggered through
`DidChangeConfigurationNotification`) is the language server side code
that parses the config sent by the client/ide. So the client/ide must
match the shape expected by the LS. This is an example why
`optOutTelemetry` did not work out of the box, and needed some changes
to be consumable by the LS.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>
Co-authored-by: Josh Pinkney <jpink@amazon.com>