feat(amazonq): add pair programming card (#7115)
## Problem
agentic chat should have a pair programming card that appears at the top
of chat

## Solution
implement it with the following behaviors:
1. If the pair programming card was dismissed in a previous session,
never show it again
2. Continue showing the pair programming card in new basic chat tabs
until you click the "X" to dismiss it. Once the "X" is pressed it will
never be shown for future tabs in the current chat-client session or
future sessions
3. If you have multiple basic chat tabs open and click X on one of them
it only closes the programmer mode card for that single tab

depends on: https://github.com/aws/language-servers/pull/1023

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.