feat(amazonq): option to show diff in reverse order (#7126)
## Problem
Sometimes we need to show the diff in reverse order (left/right
flipped).


## Solution
Add a flag to show the diff in reverse order. Right now we will always
do this in the lsp diff handler to avoid the need to create a temporary
file.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.