fix(amazonq): increase polling frequency for bearer token.  (#7123)
## Problem
During the bug bash, someone encountered the LSP attempting to use an
expired bearer token. I am unable to reproduce the issue, but can
speculate at the cause.

We refresh the bearer token on the LSP every 1 minute, but if the token
expires, and then a request is made before the next refresh it can fail.

Note: this is a temporary solution until we have LSP auth. 

## Solution
- double the frequency with which we check the token. 
- We still only send the token if it changed, meaning this won't cause
noisy request to LSP.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.