feat(amazonq): add handling for button clicks (#7125)
## Problem
With https://github.com/aws/language-servers/pull/1037, the chat client
will forward the button click event to the host client (VSC in this
case). We are then responsible for making a request to the LSP with this
button click.

## Solution
- route the button click to the LSP. 
- log any failed button click actions. 

## Verification
See https://github.com/aws/language-servers/pull/1037, for a demo of the
e2e button hookup working.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.