fix(amazonq): indicate profile needs to be selected from status bar menu
Problem:

Inline suggestion users who upgraded to the new extension version with region profile selection
need to select a profile, but are not prompted in a significant way to do so.

Solution:

Show the status bar as not connected, and once clicked they will see the top item
telling them they must select a profile to get features working.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>