fix(lsp): LSP download timeout message disappears too early
Problem:

When downloading the LSP artifacts from the manifest, there are some large ones (100+ MB).
For a slow connection this can take 1+ minutes, but our timeout for the "downloading" message
is set to disappear much earlier. So it is still downloading in the background, but the message
has disappeared and the user thinks that the download is done.

Solution:

Increase the timeout to 30 minutes, which will ensure the downloading message sticks around
while the download is still happening.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>