fix(lsp): LSP download timeout message disappears too early (#7145)
## Problem:

When downloading the LSP artifacts from the manifest, there are some
large ones (100+ MB).
For a slow connection this can take 1+ minutes, but our timeout for the
"downloading" message
is set to disappear much earlier. So it is still downloading in the
background, but the message
has disappeared and the user thinks that the download is done.

## Solution:

Increase the timeout to 30 minutes, which will ensure the downloading
message sticks around
while the download is still happening.

## Additional

In another commit this fixes a separate HTTP client bug where it would
time out a request if it took longer than 3 seconds.
This caused downloads to be aborted (separate from the download message
disappearing). Now it times out after 30 minutes.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>