fix(chat): Resend auth token on server restart
Problem:

When the server restarted due to a crash, the auth token was not
sent again.

Note that server restart is triggered automatically by the LanguageClient, I think.

Solution:

Detect when the server is restarted and manually resend the bearer token again.

Note, this solution needs to be revisited since there may be other initialization logic
that needs to run on server restart, aside from just the bearer token.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>