feat(core): add value length cap to partialClone (#7150)
## Problem
If we want to log a large json object with giant strings (think whole
files), it can make the logs difficult to read. This is especially
relevant when the underlying string values are not very important.

## Solution
- allow the string values to be truncated with `maxStringLength` option.

## Notes
I am planning to port this utility to Flare to improve the logging
experience there, and want this functionality to exist there however I
thought this could be useful here too.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.