fix(amazonq): fix enterprise users not able to sign in correctly if they have 2+ vscode instances open (#7151)
## Problem
revision of #7134, this pr aims to address the comment from the previous
PR
https://github.com/aws/aws-toolkit-vscode/pull/7134#discussion_r2056518840
to extract the logic to a shared module so the diff against 7134 is
expected to be large.


after deployment 4/21, service has a strict 1 tps throttling policy and
there was no caching for the API call previously.
It will impact users as long as they have multiple ide instances opened
as all of them will make list profile call while users attempt to sign
in.

## Solution
1. cache the api result for 60 seconds for reuse 
2. use lock to prevent multiple instances trying to call at the same
time. Only 1 will make the service call and the rest will wait until
it's done and read from the cache.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.