feat(amazonq): support multiple uri in a single diff scheme (#7167)
## Problem
The current viewDiff implementation is restricted to only allowing
comparing the current file contents. This means it's not possible to
view a diff for a change that was applied at a previous point in time.

## Solution
- Create a new `DiffContentProvider` that can support showing multiple
URIs in the same scheme. The current `ContentProvider` has a limitation
that it can only show a single URI.
- Refactor the code to use a defined type `ViewDiffMessage` instead of
`any`
- Change the logic so that diffs are stored in memory instead of a temp
file on disk

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.