fix(amazonq): Remove the 'Cancel' button on LSP download
Just remove the 'Cancel' button for now when downloading the language server.
We do not want to deal with the edge case of a user cancelling part way and then
their Q features don't work since it depends on the language server.

---

TODO: Figure out a better solution for allowing users to cancel the LS download,
but then making it easy to download again if they eventually want it. We will also
need to design it so they are aware that cancelling will prevent Q features from working.

This will require an update to the spec.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>