refactor(amazonq): Change Q LSP downloading message text (#7185)
## PROBLEM:

When installing the LSP for Q we get a message that says `"Installing
'Amazon Q' language server"`, but they want it to be slightly different
(`"Updating Amazon Q plugin"`). The issue is the base class for the lsp
installer class has a generic message that is not easy to customize.

## SOLUTION:

Allow an override downloading message to be defined in each specific
LspInstaller implementation. This message is then routed in to the
LspResolver (the thing that downloads the Lsp), and the message is
displayed here. This override message will show when the download is
happening. Everyone is happy.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>