feat(amazonq): Log attribution notice when downloading language server (#7193)
## Problem:

Legal requires us to log the attribution notice when downloading the
language server. We do not currently do this.

## Solution:

- What our message looks like: `lsp: Installing 'AmazonQ' Language
Server v0.1.0-alpha.100 to
/Users/nkomonen/Library/Caches/aws/toolkits/language-servers/AmazonQ/0.1.0-alpha.100
(Attribution notice can be found at
https://dhi0h88q3j9q6.cloudfront.net/6a0c7aa8-8b63-43a7-b14b-ee7594a29c9d/THIRD_PARTY_LICENSES)`
- This log message is in the generic part of our LSP downloader, and
will show for all LSPs downloaded. The link to the attribution notice is
already part of the manifest, so we grab it from there.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>