fix(amazonq): dev profile selection breaks if certain regions aren't … (#7203)
## Problem
Customer reported when a certain region is blocked(i.e. by cooperate
network), it causes the Q developer profile list to break, and the
customer can no longer use Q.

## Solution
Don't throw immediately when a region is unavailable. So that the user
still have access to working regions.

Only throw if all regions are failing. 


## Testing
Shared the VSIX with customer and confirmed that issue is resolved. 

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.