telemetry(amazonq): Add changed IDE diagnostics after user acceptance  (#7130)
## Problem
When user accepts suggestion from inline completion or chat, there can
be a change in the current open editor's IDE diagnostics, this can be
used as a measure for code suggestion quality. Ref:
https://github.com/aws/aws-toolkit-jetbrains/pull/5613

This change is part of the server side workspace context.

To reduce the risk of large blast radius, this change is only applied
for users who are in the experiment of server side project context (both
treatment and control), which is the Amzn idc users.

E2E verifying request id: `97a3c4d4-0c78-4329-93ae-379d2ec66646`.


## Solution
1. Add changed IDE diagnostics after user acceptance 

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.