fix(logging): misleading "settings failed" message
Problem:
Misleading log message when a setting returns `null`, even if the caller
specified `defaultValue`:

    [error] Settings (amazonQ): failed to read "workspaceIndexCacheDirPath":
    Unexpected type cast: got object, expected primitive String

`cast()` considers `null` to be "object" type, which is misleading. If
the caller passed a `defaultValue`, they have intentionally opted-in to
ignoring this situation, and *don't* want this kind of log message.

Solution:
If a setting returns `null|undefined`, skip `cast()` and return
`defaultValue` immediately. Don't log an error message.