fix(amazonq): add a more descriptive issue for firewall problems (#7251)
## Problem
If a users firewall is blocking downloading the zips or reaching the
manifest its not clear to the user what to do

## Solution
If the user:
1. doesn't have the latest version cached
2. download of the manifest failed
3. they don't have any previous language servers downloaded

it might be a firewall issue. Rather than just `Unable to find a
compatible version of the Language Server` we show a more descriptive
message to the user

<img width="454" alt="Screenshot 2025-05-07 at 2 48 39 PM"
src="https://github.com/user-attachments/assets/48053a0f-7dff-4ba2-8b7d-65769b30a56f"
/>


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.