telemetry(amazonq): add an error code for when node validation fails.  (#7254)
## Problem
These errors show up in telemetry with reason `Error`. All of the
`reasonDesc` are different since they include unique PID values. This
makes it hard to count/quantify how often this is happening.

## Solution
- attach a unique error code to these errors. For when node can't be run
we have `FailedToRunNode`, and when the language server can't be started
we have `FailedToStartLanguageServer`.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.