test(amazonq): reduce flakiness by avoiding unnecessary fs operations.  (#7259)
## Problem
This test has been very flaky, and constantly failing CI
(https://github.com/aws/aws-toolkit-vscode/issues/7187).
This test is run for almost 40 different cases, and does the following:
- create a text document
- write it to the filesystem. 
- run the check on the text document. 

One of the core issues is that when we write the text document to the
filesystem, we create a new test workspace folder for each case. This
involves creating a directory with a random id for each of the almost 40
different cases.

These excessive file system operations could be leading to the flakiness
in CI.

## Solution
- reuse the same test workspace folder across all test cases. This
should cut the file systems operations in half.

## Future Work
- If this doesn't reduce flakiness, we could avoid writing the text
document to the fs since its not needed by the underlying test, but this
involves mocking the text document which is undesirable imo.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.