feat(amazonq): re-add activeState/line trackers (#7257)
## Problem
- "Amazon Q is generating ..." does not show with the lsp mode

## Solution
- re-add the line tracker with tests
- re-implement the activeState tracker

## Notes
In master the active state tracker decides whether or not to show
"Amazon Q is generating ..." by the following:
- When a change is made, the auto trigger decides whether or not to
start a recommendation request. When a recommendation requests
eventually starts, an event is sent to the active state tracker to tell
it to start showing the "Amazon Q is generating ..." message. When the
first recommendation starts loading and the results are shown to the
user another event is sent telling it to hide the message. It de-bounces
this message showing every 1000ms so that the message is not constantly
toggling on/off

In this implementation its slightly different:
- VSCode decides when to trigger the inline completion through their
inline completion provider. From here we show the "Amazon Q is
generating ... " message until the first recommendation is received from
the language server and shown to the user. It still de-bounces this
message every 1000ms so that users aren't shown it too often


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.