fix: Show "Must select Developer Profile" message
Problem:

THe show developer profile message was appearing at the wrong time.
It was due to the order that the code executed. This was because when we
called to check if the profiles were restored from cache, the logic that
restored them had not actually run yet.

Solution:

Move the logic that decides to show the Needs Profile message, to after the code that
restores the Profile.

It is also only called once since we only want to show the message on startup.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>